==> data_analysis/3_average_data.ipynb <==
## Importing Libraries

# Importing relevant libraries
import pandas as pd
import numpy as np
from pathlib import Path
import os
import matplotlib
import matplotlib.pyplot as plt
import math

## Getting All Data

# Retrieving filepath to all data folders
%store -r fp_folder

# Check if the file exists before attempting to delete it
if os.path.exists(fp_folder + 'averaged_results.csv'):
    # Attempt to delete the file
    try:
        os.remove(fp_folder + 'averaged_results.csv')
        print(f"File '{fp_folder + 'averaged_results.csv'}' has been deleted successfully.")
    except OSError as e:
        print(f"Error deleting file '{fp_folder + 'averaged_results.csv'}': {e}")
else:
    print(f"File '{fp_folder + 'averaged_results.csv'}' does not exist.")

sub_folders = []
for i in os.listdir(fp_folder):
    if os.path.isdir(os.path.join(fp_folder, i)) and (i.startswith('BNC') or i.startswith('NAV')):
        sub_folders.append(i)

num_subjects = len(sub_folders)
sub_folders = sorted(sub_folders)
print(sub_folders)

df_final = pd.DataFrame({
        'Participant': [],
        'Block_Num': [],
        'Total_Time': [],
        'Orientation_Time': [],
        'Navigation_Time': [],
        'Distance': [],
        'Speed': [],
        'Mean_Dwell': [],
        'Teleportations': [],
        'Mean_Teleport_Distance': []
    })

# Iterating through participant folders and conducting analyses
for PID in sub_folders:
    
    # Filepath to each participant folder
    fp_PID = fp_folder + PID

    for i in range(1,4):
        fp_results = fp_PID +'/b'+str(i)+'_results.csv'
        data = pd.read_csv(fp_results)
        total_time = data['Total_Time'].mean()
        orientation_time = data['Orientation_Time'].mean()
        nav_time = data['Navigation_Time'].mean()
        distance = data['Distance'].mean()
        speed = data['Speed'].mean()
        mean_dt = data['Mean_Dwell'].mean()
        teleports = data['Teleportations'].mean()
        mean_teleport_distance = data['Mean_Teleport_Distance'].mean()
        new_row = {
                    'Participant': PID,
                    'Block_Num': i,
                    'Total_Time': total_time,
                    'Orientation_Time': orientation_time,
                    'Navigation_Time': nav_time,
                    'Distance': distance,
                    'Speed': speed,
                    'Mean_Dwell': mean_dt,
                    'Teleportations': teleports,
                    'Mean_Teleport_Distance': mean_teleport_distance
                }
        df_final.loc[len(df_final)] = new_row

df_final.to_csv(fp_folder + 'averaged_results.csv', index=True, header = True)
